requirements.txt
pymemcache
pillow